health_system/requirements.txt
Django~=5.0.1
Pillow
django-crispy-forms
djangorestframework
markdown       
django-filter  
django-rest-swagger
bootstrap4
crispy-bootstrap4 
django-leaflet
django-redis
reportlab
virtualenvwrapper
drf-yasg
google-cloud
google-cloud-vision
firebase-admin